Move functions out of deltagenerator
